feat: Implement MessageProcessor module for managing conversation turns
- Added MessageProcessor module with core functionality for orchestrating conversation turns.
- Introduced components: EventHandler, TaskManager, and data models for managing turn states and events.
- Implemented async task tracking, event streaming, and deterministic cleanup.
- Created README documentation detailing module structure, usage, and design principles.
- Added constants for configuration and utility functions for task management.
- Updated tests to ensure integration with the new MessageProcessor functionality.
