+++ src/services/websocket_service/message_processor/__init__.py
@@ -1,0 +1,22 @@
+"""MessageProcessor module for managing conversation turns and event streaming.
+
+This module provides the core orchestration for conversation turns, managing
+async tasks, coordinating AgentService streaming events, and guaranteeing
+deterministic cleanup.
+"""
+
+from .constants import INTERRUPT_WAIT_TIMEOUT, TOKEN_QUEUE_SENTINEL
+from .event_handlers import EventHandler
+from .models import ConversationTurn, TurnStatus
+from .processor import MessageProcessor
+from .task_manager import TaskManager
+
+__all__ = [
+    "MessageProcessor",
+    "ConversationTurn",
+    "TurnStatus",
+    "EventHandler",
+    "TaskManager",
+    "TOKEN_QUEUE_SENTINEL",
+    "INTERRUPT_WAIT_TIMEOUT",
+]
