+++ src/services/websocket_service/message_processor/constants.py
@@ -1,0 +1,7 @@
+"""Constants for MessageProcessor."""
+
+# Sentinel object to signal end of token stream
+TOKEN_QUEUE_SENTINEL = object()
+
+# Timeout for interrupt operations
+INTERRUPT_WAIT_TIMEOUT = 1.0
