+++ src/services/websocket_service/message_processor/processor.py
@@ -9,58 +9,15 @@ from __future__ import annotations
 
 import asyncio
 import time
-from contextlib import suppress
-from dataclasses import dataclass, field
-from enum import Enum
 from typing import Any, AsyncGenerator, AsyncIterator, Dict, List, Optional, Set
 from uuid import UUID, uuid4
 
 from loguru import logger
 
-from .text_processors import TextChunkProcessor, TTSTextProcessor
-
-_TOKEN_QUEUE_SENTINEL = object()
-_INTERRUPT_WAIT_TIMEOUT = 1.0
-
-
-class TurnStatus(Enum):
-    """Status of a conversation turn."""
-
-    PENDING = "pending"
-    PROCESSING = "processing"
-    COMPLETED = "completed"
-    INTERRUPTED = "interrupted"
-    FAILED = "failed"
-
-
-@dataclass
-class ConversationTurn:
-    """Represents a single conversation turn."""
-
-    turn_id: str
-    user_message: str
-    conversation_id: str
-    status: TurnStatus = TurnStatus.PENDING
-    created_at: float = field(default_factory=time.time)
-    updated_at: float = field(default_factory=time.time)
-    metadata: Dict[str, Any] = field(default_factory=dict)
-    tasks: Set[asyncio.Task] = field(default_factory=set)
-    response_content: str = ""
-    error_message: Optional[str] = None
-    event_queue: Optional[asyncio.Queue] = None
-    token_queue: Optional[asyncio.Queue] = None
-    token_consumer_task: Optional[asyncio.Task] = None
-    token_stream_closed: bool = False
-    chunk_processor: Optional[TextChunkProcessor] = None
-    tts_processor: Optional[TTSTextProcessor] = None
-
-    def update_status(self, status: TurnStatus, error_message: Optional[str] = None):
-        """Update turn status and timestamp."""
-        self.status = status
-        self.updated_at = time.time()
-        if error_message:
-            self.error_message = error_message
-        logger.debug(f"Turn {self.turn_id} status updated to {status.value}")
+from .constants import INTERRUPT_WAIT_TIMEOUT
+from .event_handlers import EventHandler
+from .models import ConversationTurn, TurnStatus
+from .task_manager import TaskManager
 
 
 class MessageProcessor:
@@ -95,6 +52,10 @@ class MessageProcessor:
         self._cleanup_lock = asyncio.Lock()
         self._current_turn_id: Optional[str] = None
         self._cleaned_turns: Set[str] = set()
+
+        # Initialize helper components
+        self._event_handler = EventHandler(self)
+        self._task_manager = TaskManager(self)
 
         logger.info(
             "MessageProcessor initialized for connection %s, user %s",
@@ -137,8 +98,6 @@ class MessageProcessor:
             turn.event_queue = asyncio.Queue(maxsize=self.queue_maxsize)
             turn.token_queue = asyncio.Queue(maxsize=self.queue_maxsize)
             turn.token_stream_closed = False
-            turn.chunk_processor = TextChunkProcessor()
-            turn.tts_processor = TTSTextProcessor()
             self.turns[turn_id] = turn
             self.active_turns.add(turn_id)
             self.total_turns += 1
@@ -151,14 +110,14 @@ class MessageProcessor:
                 conversation_id,
             )
 
-        self._ensure_token_consumer(turn_id)
+        self._task_manager.ensure_token_consumer(turn_id)
 
         if agent_stream is not None:
             producer_task = asyncio.create_task(
-                self._produce_agent_events(turn_id, agent_stream),
+                self._event_handler.produce_agent_events(turn_id, agent_stream),
                 name=f"message-processor-producer-{turn_id}",
             )
-            self._track_task(turn_id, producer_task)
+            self._task_manager.track_task(turn_id, producer_task)
 
         return turn_id
 
@@ -200,9 +159,7 @@ class MessageProcessor:
             logger.warning("Turn %s not found for task addition", turn_id)
             return False
 
-        turn.tasks.add(task)
-        self.active_tasks.add(task)
-        task.add_done_callback(self._build_task_cleanup(turn_id))
+        self._task_manager.track_task(turn_id, task)
 
         logger.debug(
             "Added task to turn %s, total tracked tasks: %d",
@@ -336,13 +293,13 @@ class MessageProcessor:
         if previous_status != TurnStatus.INTERRUPTED:
             self.total_interrupted += 1
 
-        await self._signal_token_stream_closed(target_turn_id)
-        await self._wait_for_token_queue(target_turn_id)
-        await self._cancel_turn_tasks(target_turn_id)
+        await self._event_handler._signal_token_stream_closed(target_turn_id)
+        await self._event_handler._wait_for_token_queue(target_turn_id)
+        await self._task_manager.cancel_turn_tasks(target_turn_id)
 
         queue = self.get_event_queue(target_turn_id)
         if queue:
-            drained = self._drain_event_queue(target_turn_id)
+            drained = self._task_manager.drain_event_queue(target_turn_id)
             if drained:
                 logger.debug(
                     "Drained %d queued events before interrupting turn %s",
@@ -365,7 +322,7 @@ class MessageProcessor:
                 await queue.put(interrupt_event)
 
             try:
-                await asyncio.wait_for(queue.join(), timeout=_INTERRUPT_WAIT_TIMEOUT)
+                await asyncio.wait_for(queue.join(), timeout=INTERRUPT_WAIT_TIMEOUT)
             except asyncio.TimeoutError:
                 logger.debug(
                     "Timed out waiting for interrupt event delivery for turn %s",
@@ -394,7 +351,7 @@ class MessageProcessor:
                 return
 
             if turn.token_queue:
-                await self._signal_token_stream_closed(turn_id)
+                await self._event_handler._signal_token_stream_closed(turn_id)
 
             current_task = asyncio.current_task()
             tasks_to_cancel: List[asyncio.Task] = []
@@ -466,13 +423,13 @@ class MessageProcessor:
         if not turn:
             raise ValueError(f"Unknown turn: {turn_id}")
 
-        self._ensure_token_consumer(turn_id)
+        self._task_manager.ensure_token_consumer(turn_id)
 
         producer_task = asyncio.create_task(
-            self._produce_agent_events(turn_id, agent_stream),
+            self._event_handler.produce_agent_events(turn_id, agent_stream),
             name=f"message-processor-producer-{turn_id}",
         )
-        self._track_task(turn_id, producer_task)
+        self._task_manager.track_task(turn_id, producer_task)
 
     async def get_turn(self, turn_id: str) -> Optional[ConversationTurn]:
         """Get a specific conversation turn."""
@@ -564,64 +521,6 @@ class MessageProcessor:
         else:
             await self.cleanup_completed_turns(0)
 
-    async def _produce_agent_events(
-        self, turn_id: str, agent_stream: AsyncIterator[Dict[str, Any]]
-    ) -> None:
-        """Consume AgentService events and forward them to the queue."""
-
-        try:
-            async for raw_event in agent_stream:
-                event = self._normalize_event(turn_id, raw_event)
-                event_type = event.get("type")
-
-                if event_type == "stream_token":
-                    await self._put_token_event(turn_id, event)
-                    continue
-
-                if event_type == "stream_start":
-                    await self._put_event(turn_id, event)
-                    await self.update_turn_status(turn_id, TurnStatus.PROCESSING)
-                    continue
-
-                if event_type == "stream_end":
-                    await self._signal_token_stream_closed(turn_id)
-                    await self._wait_for_token_queue(turn_id)
-                    await self._put_event(turn_id, event)
-                    await self.complete_turn(turn_id)
-                    continue
-
-                if event_type == "error":
-                    await self._signal_token_stream_closed(turn_id)
-                    await self._wait_for_token_queue(turn_id)
-                    await self._put_event(turn_id, event)
-                    await self.fail_turn(turn_id, event.get("error", "Unknown error"))
-                    continue
-
-                logger.debug(
-                    "Dropping agent event %s from client stream for turn %s",
-                    event_type,
-                    turn_id,
-                )
-
-        except asyncio.CancelledError:
-            logger.debug("Producer cancelled for turn %s", turn_id)
-            raise
-        except Exception as exc:  # pragma: no cover - defensive  # noqa: BLE001
-            await self.fail_turn(turn_id, str(exc))
-            await self._signal_token_stream_closed(turn_id)
-            await self._wait_for_token_queue(turn_id)
-            await self._put_event(
-                turn_id,
-                {
-                    "type": "error",
-                    "error": str(exc),
-                    "turn_id": turn_id,
-                },
-            )
-        finally:
-            await self._signal_token_stream_closed(turn_id)
-            logger.debug("Producer finished for turn %s", turn_id)
-
     async def stream_events(
         self, turn_id: Optional[str] = None
     ) -> AsyncGenerator[Dict[str, Any], None]:
@@ -666,258 +565,6 @@ class MessageProcessor:
             queue.qsize(),
         )
 
-    async def _cancel_turn_tasks(
-        self,
-        turn_id: str,
-        *,
-        timeout: float = _INTERRUPT_WAIT_TIMEOUT,
-    ) -> None:
-        """Cancel background tasks associated with a turn and wait briefly."""
-
-        turn = self.turns.get(turn_id)
-        if not turn or not turn.tasks:
-            return
-
-        current_task = asyncio.current_task()
-        tasks_to_cancel: List[asyncio.Task] = []
-
-        for task in list(turn.tasks):
-            if not isinstance(task, asyncio.Task):
-                continue
-            if task.done() or task is current_task:
-                continue
-            task.cancel()
-            tasks_to_cancel.append(task)
-
-        if not tasks_to_cancel:
-            return
-
-        done, pending = await asyncio.wait(
-            tasks_to_cancel,
-            timeout=timeout,
-            return_when=asyncio.ALL_COMPLETED,
-        )
-
-        for task in done:
-            with suppress(asyncio.CancelledError, Exception):
-                task.result()
-
-        if pending:
-            logger.debug(
-                "Timed out waiting for %d tasks to cancel for turn %s",
-                len(pending),
-                turn_id,
-            )
-
-    def _drain_event_queue(self, turn_id: str) -> int:
-        """Remove any pending events from the client queue."""
-
-        queue = self.get_event_queue(turn_id)
-        if not queue:
-            return 0
-
-        drained = 0
-        while True:
-            try:
-                queue.get_nowait()
-                queue.task_done()
-                drained += 1
-            except asyncio.QueueEmpty:
-                break
-
-        return drained
-
-    def _ensure_token_consumer(self, turn_id: str) -> None:
-        """Ensure the token consumer task exists for the given turn."""
-
-        turn = self.turns.get(turn_id)
-        if not turn:
-            return
-
-        if turn.token_queue is None:
-            turn.token_queue = asyncio.Queue(maxsize=self.queue_maxsize)
-            turn.token_stream_closed = False
-
-        existing_consumer = turn.token_consumer_task
-        if existing_consumer and not existing_consumer.done():
-            return
-
-        consumer_task = asyncio.create_task(
-            self._consume_token_events(turn_id),
-            name=f"message-processor-consumer-{turn_id}",
-        )
-        turn.token_consumer_task = consumer_task
-        self._track_task(turn_id, consumer_task)
-
-    async def _put_token_event(self, turn_id: str, event: Dict[str, Any]) -> None:
-        """Send a token event to the internal token queue."""
-
-        turn = self.turns.get(turn_id)
-        queue = turn.token_queue if turn else None
-        if not queue:
-            logger.debug(
-                "Dropping token event for turn %s due to missing queue", turn_id
-            )
-            return
-
-        await queue.put(event)
-        logger.debug(
-            "Queued token event for turn %s (queue size=%d)",
-            turn_id,
-            queue.qsize(),
-        )
-
-    async def _consume_token_events(self, turn_id: str) -> None:
-        """Consume token events and emit TTS ready chunks."""
-
-        turn = self.turns.get(turn_id)
-        if not turn:
-            return
-
-        queue = turn.token_queue
-        if queue is None:
-            return
-
-        try:
-            while True:
-                token_event = await queue.get()
-                try:
-                    if token_event is _TOKEN_QUEUE_SENTINEL:
-                        await self._flush_tts_buffer(turn_id)
-                        break
-
-                    await self._process_token_event(turn_id, token_event)
-                finally:
-                    queue.task_done()
-        except asyncio.CancelledError:
-            logger.debug("Token consumer cancelled for turn %s", turn_id)
-            raise
-        except Exception as exc:  # pragma: no cover - defensive  # noqa: BLE001
-            logger.error("Error consuming token events for turn %s: %s", turn_id, exc)
-        finally:
-            turn = self.turns.get(turn_id)
-            if turn:
-                turn.token_stream_closed = True
-
-    async def _process_token_event(
-        self,
-        turn_id: str,
-        token_event: Dict[str, Any],
-    ) -> None:
-        """Transform a single token event into zero or more TTS events."""
-
-        turn = self.turns.get(turn_id)
-        if not turn:
-            logger.debug("Token event received for unknown turn %s", turn_id)
-            return
-
-        if not turn.chunk_processor:
-            turn.chunk_processor = TextChunkProcessor()
-        if not turn.tts_processor:
-            turn.tts_processor = TTSTextProcessor()
-
-        chunk = token_event.get("data") or token_event.get("chunk")
-        if not chunk:
-            logger.debug("Token event missing chunk data: %s", token_event)
-            return
-
-        for sentence in turn.chunk_processor.process(chunk):
-            processed = turn.tts_processor.process(sentence)
-            text = processed.filtered_text
-            if not text or not any(char.isalnum() for char in text):
-                continue
-            tts_event = self._build_tts_event(
-                turn_id,
-                token_event,
-                text,
-                processed.emotion_tag,
-            )
-            await self._put_event(turn_id, tts_event)
-
-    async def _flush_tts_buffer(self, turn_id: str) -> None:
-        """Flush any remaining buffered text for a turn."""
-
-        turn = self.turns.get(turn_id)
-        if not turn or not turn.chunk_processor or not turn.tts_processor:
-            return
-
-        remainder = turn.chunk_processor.flush()
-        if not remainder:
-            return
-
-        processed = turn.tts_processor.process(remainder)
-        text = processed.filtered_text
-        if not text or not any(char.isalnum() for char in text):
-            return
-
-        tts_event = self._build_tts_event(
-            turn_id,
-            {},
-            text,
-            processed.emotion_tag,
-        )
-        await self._put_event(turn_id, tts_event)
-
-    def _build_tts_event(
-        self,
-        turn_id: str,
-        base_event: Dict[str, Any],
-        chunk: str,
-        emotion: Optional[str] = None,
-    ) -> Dict[str, Any]:
-        """Construct a tts_ready_chunk event with optional emotion metadata."""
-
-        event = self._normalize_event(turn_id, base_event)
-        tts_event = {
-            key: value
-            for key, value in event.items()
-            if key not in {"type", "data", "chunk"}
-        }
-        tts_event["type"] = "tts_ready_chunk"
-        tts_event["chunk"] = chunk
-        if emotion:
-            tts_event["emotion"] = emotion
-        return tts_event
-
-    async def _signal_token_stream_closed(self, turn_id: str) -> None:
-        """Notify the token consumer that no more tokens will arrive."""
-
-        turn = self.turns.get(turn_id)
-        queue = turn.token_queue if turn else None
-        if not queue or turn.token_stream_closed:
-            return
-
-        try:
-            queue.put_nowait(_TOKEN_QUEUE_SENTINEL)
-        except asyncio.QueueFull:
-            await queue.put(_TOKEN_QUEUE_SENTINEL)
-
-        turn.token_stream_closed = True
-
-    async def _wait_for_token_queue(self, turn_id: str) -> None:
-        """Wait for the token queue to drain pending items."""
-
-        turn = self.turns.get(turn_id)
-        queue = turn.token_queue if turn else None
-        if not queue:
-            return
-
-        try:
-            await asyncio.wait_for(queue.join(), timeout=_INTERRUPT_WAIT_TIMEOUT)
-        except asyncio.TimeoutError:
-            logger.debug(
-                "Timed out waiting for token queue join for turn %s",
-                turn_id,
-            )
-        except asyncio.CancelledError:  # pragma: no cover - defensive
-            raise
-        except Exception as exc:  # pragma: no cover - defensive  # noqa: BLE001
-            logger.debug(
-                "Error waiting for token queue for turn %s: %s",
-                turn_id,
-                exc,
-            )
-
     async def _delayed_cleanup(self, delay: float):
         """Delayed cleanup of resources."""
 
@@ -945,28 +592,6 @@ class MessageProcessor:
         normalized.setdefault("connection_id", str(self.connection_id))
         normalized.setdefault("user_id", self.user_id)
         return normalized
-
-    def _track_task(self, turn_id: str, task: asyncio.Task) -> None:
-        """Track lifecycle of a background task."""
-
-        turn = self.turns.get(turn_id)
-        if not turn:
-            return
-
-        turn.tasks.add(task)
-        self.active_tasks.add(task)
-        task.add_done_callback(self._build_task_cleanup(turn_id))
-
-    def _build_task_cleanup(self, turn_id: str):
-        """Create a callback that removes completed tasks from tracking."""
-
-        def _callback(task: asyncio.Task) -> None:
-            self.active_tasks.discard(task)
-            turn = self.turns.get(turn_id)
-            if turn:
-                turn.tasks.discard(task)
-
-        return _callback
 
     async def _default_agent_stream(
         self, turn_id: str, conversation_id: str, user_input: str
